修改&重構
將原roomtype表格拆開
並根據其跟改的欄位名稱重新修改java

html加上hover的效果

+++ common/src/main/java/tw/edu/ntut/sce/java18/common/dao/RoomDao.java
@@ -6,7 +6,7 @@ import tw.edu.ntut.sce.java18.common.model.RoomBean;
 
 public interface RoomDao {
 
-  Map<Integer, Long> getAvailableFloorCount(String type);
+  Map<Integer, Integer> getAvailableFloorCount(String type);
 
   ArrayList<String> getEmptyRoom();
